requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
rapidfuzz>=3.9.0
Jinja2>=3.1.2
matplotlib>=3.7.2
fpdf2>=2.7.6
